Update folder structure to match pensions and income_and_assets modules
