+++ modules/burials/lib/burials/engine.rb
@@ -19,11 +19,11 @@ module Burials
     initializer 'burials.pdf_fill.register_form' do |app|
       app.config.to_prepare do
         require 'pdf_fill/filler'
-        require 'burials/pdf_fill/forms/va21p530ez'
+        require 'burials/pdf_fill/va21p530ez'
 
         # Register our Burial Pdf Fill form
-        ::PdfFill::Filler.register_form(Burials::PdfFill::Forms::Va21p530ez::FORM_ID,
-                                        Burials::PdfFill::Forms::Va21p530ez)
+        ::PdfFill::Filler.register_form(Burials::PdfFill::Va21p530ez::FORM_ID,
+                                        Burials::PdfFill::Va21p530ez)
       end
     end
